NewWizard: Refactor and break apart to ease testing
