+++ src/new/NewWizard.js
@@ -1,46 +1,16 @@
 // @flow
-import React, { type Node, Component, createContext } from 'react';
+import React, { type Node, Component } from 'react';
 import StepWrapper, { type StepProps } from './StepWrapper';
+import { WizardContext, type ContextType } from './WizardContext';
+import { ConsumeValidator } from './ConsumeValidator';
 
 type WizardProps = {
   children: Node,
 };
 
-type StepType = {
-  name: string,
-};
-
-type ContextType = {|
-  stepList: Array<StepType>,
-|};
-
-type ConsumerProps = {
-  children: ContextType => Node,
-};
-
 type State = {
   context: ContextType,
 };
-
-const WizardContext = createContext();
-
-function ConsumeValidator(props: ConsumerProps) {
-  return (
-    <WizardContext.Consumer>
-      {context => {
-        if (!context) {
-          throw new Error(
-            '[React Losen]: Step and Controls must be rendered within the <Wizard /> component',
-          );
-        }
-        if (typeof props.children === 'function') {
-          return props.children(context);
-        }
-        return null;
-      }}
-    </WizardContext.Consumer>
-  );
-}
 
 class NewWizard extends Component<WizardProps, State> {
   static Step = (props: StepProps) => (
