+++ src/new/WizardContext.js
@@ -1,0 +1,12 @@
+// @flow
+import { createContext } from 'react';
+
+export type StepType = {
+  name: string,
+};
+
+export type ContextType = ?{|
+  stepList: Array<StepType>,
+|};
+
+export const WizardContext = createContext<ContextType>(null);
